requirements.txt
# Document ingestion dependencies
pypdf>=3.17.0
python-docx>=1.1.0

# Text chunking dependencies
langchain>=0.1.0
langchain-text-splitters>=0.0.1

# Embedding dependencies
sentence-transformers>=2.2.0
torch>=2.0.0

# Vector storage dependencies
faiss-cpu>=1.7.4

# LLM dependencies
ollama>=0.1.0

# API dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0

